reorganize configs
