+++ nemo/collections/asr/models/configs/ctc_models_config.py
@@ -12,51 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import Any, Dict, List, Optional
 
-from omegaconf import MISSING, OmegaConf
+from omegaconf import MISSING
 
+from nemo.collections.asr.modules.audio_preprocessing import (
+    AudioToMelSpectrogramPreprocessorConfig,
+    SpectrogramAugmentationConfig,
+)
+from nemo.collections.asr.modules.conv_asr import ConvASRDecoderConfig, ConvASREncoderConfig
 from nemo.core.config import modelPT as model_cfg
-
-
-@dataclass
-class AudioToMelSpectrogramPreprocessorConfig:
-    _target_: str = "nemo.collections.asr.modules.AudioToMelSpectrogramPreprocessor"
-    sample_rate: int = 16000
-    window_size: float = 0.02
-    window_stride: float = 0.01
-    n_window_size: Optional[int] = None
-    n_window_stride: Optional[int] = None
-    window: str = "hann"
-    normalize: str = "per_feature"
-    n_fft: Optional[int] = None
-    preemph: float = 0.97
-    features: int = 64
-    lowfreq: int = 0
-    highfreq: Optional[int] = None
-    log: bool = True
-    log_zero_guard_type: str = "add"
-    log_zero_guard_value: float = 2 ** -24
-    dither: float = 1e-5
-    pad_to: int = 16
-    frame_splicing: int = 1
-    stft_exact_pad: bool = False
-    stft_conv: bool = False
-    pad_value: int = 0
-    mag_power: float = 2.0
-
-
-@dataclass
-class SpecAugmentConfig:
-    _target_: str = "nemo.collections.asr.modules.SpectrogramAugmentation"
-    freq_masks: int = 0
-    time_masks: int = 0
-    freq_width: int = 0
-    time_width: Optional[Any] = 0
-    rect_masks: int = 0
-    rect_time: int = 0
-    rect_freq: int = 0
 
 
 @dataclass
@@ -87,53 +53,6 @@ class EncDecCTCDatasetConfig(model_cfg.DatasetConfig):
 
 
 @dataclass
-class JasperEncoderConfig:
-    filters: int = MISSING
-    repeat: int = MISSING
-    kernel: List[int] = MISSING
-    stride: List[int] = MISSING
-    dilation: List[int] = MISSING
-    dropout: float = MISSING
-    residual: bool = MISSING
-
-    # Optional arguments
-    groups: int = 1
-    separable: bool = False
-    heads: int = -1
-    residual_mode: str = "add"
-    residual_dense: bool = False
-    se: bool = False
-    se_reduction_ratio: int = 8
-    se_context_size: int = -1
-    se_interpolation_mode: str = 'nearest'
-    kernel_size_factor: float = 1.0
-    stride_last: bool = False
-
-
-@dataclass
-class ConvASREncoderConfig:
-    _target_: str = 'nemo.collections.asr.modules.ConvASREncoder'
-    jasper: Optional[JasperEncoderConfig] = field(default_factory=list)
-    activation: str = MISSING
-    feat_in: int = MISSING
-    normalization_mode: str = "batch"
-    residual_mode: str = "add"
-    norm_groups: int = -1
-    conv_mask: bool = True
-    frame_splicing: int = 1
-    init_mode: str = "xavier_uniform"
-
-
-@dataclass
-class ConvASRDecoderConfig:
-    _target_: str = 'nemo.collections.asr.modules.ConvASRDecoder'
-    feat_in: int = MISSING
-    num_classes: int = MISSING
-    init_mode: str = "xavier_uniform"
-    vocabulary: Optional[List[str]] = field(default_factory=list)
-
-
-@dataclass
 class EncDecCTCConfig(model_cfg.ModelConfig):
     # Model global arguments
     sample_rate: int = 16000
@@ -152,7 +71,7 @@ class EncDecCTCConfig(model_cfg.ModelConfig):
 
     # Model component configs
     preprocessor: AudioToMelSpectrogramPreprocessorConfig = AudioToMelSpectrogramPreprocessorConfig()
-    spec_augment: Optional[SpecAugmentConfig] = SpecAugmentConfig()
+    spec_augment: Optional[SpectrogramAugmentationConfig] = SpectrogramAugmentationConfig()
     encoder: Any = ConvASREncoderConfig()
     decoder: Any = ConvASRDecoderConfig()
